Refactor : body_type 값 불러오는 로직 수정
