+++ app/controllers/user_controller.py
@@ -1,4 +1,4 @@
-from app.services.squatService.squatService import process_squat
+from app.services.squatService.squat_service import process_squat
 from app.services.sholderPressService.sholderPressService import process_dumbbell_shoulderPress
 from app.services.user_info_service import save_user_and_body_data_and_body_type
 from flask import Blueprint, jsonify, request
