+++ app/shared/global_state.py
@@ -11,7 +11,7 @@ accel_seq_buffer = deque(maxlen=30)
 fall_detected = False
 is_first = True  # 현재 클라이언트로 전달받은 데이터가 맨 처음 데이터인지 확인
 distances = {}  # 유저의 각 landmark 사이의 거리
-current_user_body_type = None  # 현재 사용자의 체형 타입
+current_user_body_type = None  # 추가: 전체 체형 정보를 저장할 변수
 client_sid = None  # 현재 연결된 클라이언트의 세션 ID
 
 # 운동 카운터 인스턴스 - 전역으로 관리
@@ -49,11 +49,10 @@ def reset_globals():
     # 뼈 길이 초기화
     distances = {}
 
-    # 사용자 body_type 초기화
+    # 추가: 전체 체형 정보 초기화
     current_user_body_type = None
 
     print(f'현재 개수 : {press_counter.count}')
-
 
     # 클라이언트 SID 초기화
     client_sid = None
@@ -64,7 +63,6 @@ def reset_globals():
     init_counters()
 
     print(f'{press_counter.count}')
-
 
     print('❌❌❌뼈 길이 데이터 빈 배열로 초기화 완료❌❌❌')
     print("🌀 전역 상태가 초기화되었습니다.")
